webservice-ssp/requirements.txt
flask
PyMySQL
PyYAML
Flask-SQLAlchemy
statsmodels
scipy
pandas